Upgrade node version in packages.json and the Dockerfile [ci all]
Using version 220 of `jest-localstorage-mock` required that we upgrade
our version of NodeJS from our previous minimum of 6.11.3. To support
that, this commit switches from the deprecated
`java:openjdk-8-jdk-alpine` docker image to the newer
`openjdk:8-jdk-alpine` image. This image is running Alpine version 3.8 which
has NodeJS 8.11.4 in it's repository. This also bumps the minimum
version of node in the packages.json file.